Allow plugins to extend the login form with password alternatives
The previous code provides no easy/elegant way for an extension to extend the login form. In my use case, I am developing a plugin which will allow the user to scan a visible code with their smartphone, as a non-compulsory alternative to a password.

This patch provides a filter which allows an extension to instruction Affiliate-WP that it does not need to verify the password (which would be used in the case that the extension has performed its own verification).